Add test for updateGrantsGrantees download function
